### 2024-01-23 #6266 @editorialbot
[PRE REVIEW]: PhyloX: A Python package for complete phylogenetic network workflows
**Submitting author:** @RemieJanssen (<a href="http://orcid.org/0000-0002-5192-1470">Remie Janssen</a>)
**Repository:** https://github.com/RemieJanssen/phylox
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
