### 2024-02-22 #6266 @fboehm
@SimonGreenhill @ulfaslak @shaunpwilkinson @azais @bgyori - Would you be willing to review this submission for JOSS? Thank you!
