### 2024-03-09 RemieJanssen/PhyloX#42 @abhishektiwari
Example code is not working
**Describe the bug**
Running example code [here](https://phylox.readthedocs.io/en/latest/index.html)

```py
from phylox.generators.randomTC import generate_network_random_tree_child_sequence
from phylox.generators.mcmc import sample_mcmc_networks
from phylox.classes import is_orchard
from phylox.rearrangement.move import MoveType

# Generate an arbitrary orchard network with 10 leaves and 5 reticulations
start_network = generate_network_random_tree_child_sequence(10, 5)
# Generate 1000 orchard networks with 10 leaves and 5 reticulations
sampled_networks = sample_mcmc_networks(
   start_network,
   {MoveType.TAIL: 0.5, MoveType.HEAD: 0.5},
   number_of_samples=100,
   burn_in=100,
   restriction_map=is_orchard,
   seed=1234,
   add_root_if_necessary=True,
)
# Write the sampled networks to a file
with open("sampled_networks.nwk", "w") as f:
   for network in sampled_networks:
      f.write(network.newick() + "\n")
```

**Error**
```
Traceback (most recent call last):
  File "/home/ubuntu/Home/joss/phylox/test.py", line 9, in <module>
    sampled_networks = sample_mcmc_networks(
  File "/home/ubuntu/.pyenv/versions/phylox/lib/python3.9/site-packages/networkx/utils/decorators.py", line 770, in func
    return argmap._lazy_compile(__wrapper)(*args, **kwargs)
  File "<class 'networkx.utils.decorators.argmap'> compilation 20", line 4, in argmap_sample_mcmc_networks_17
  File "/home/ubuntu/.pyenv/versions/phylox/lib/python3.9/site-packages/phylox/generators/mcmc/base.py", line 173, in sample_mcmc_networks
    if not (restriction_map is None or restriction_map(result_network)):
  File "/home/ubuntu/.pyenv/versions/phylox/lib/python3.9/site-packages/phylox/classes/dinetwork.py", line 90, in is_orchard
    leaves.remove(pair[0])
KeyError: 9
```

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6427)
