### 2024-03-09 RemieJanssen/PhyloX#43 @abhishektiwari
Typos in paper.md
1. `Summary` section : `NetworkX`not `networkX`
2.  `Statement of Need` section:  `phylogenetic` not `pylogenetic`
3.  `PhyloNetwork` section:  `rearrangement` not `rearragement`

For [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6427)
