### 2024-03-09 #6427 @abhishektiwari
@RemieJanssen  Please see my initial comments.

[Example code](https://github.com/RemieJanssen/PhyloX/issues/42) on documentation landing page did not run for me. Otherwise I was able run test scripts and create my own scripts without any issues.

A few minor typos in paper reported [here](https://github.com/RemieJanssen/PhyloX/issues/43). I also think first line of paper's `Summary` section as well other places describing `PhyloX` will read just fine without `with tools`, i.e. "_PhyloX is a Python package for generating, manipulating, and analyzing phylogenetic networks._"
