### 2024-03-23 #6427 @bgyori
I have now completed my initial review, see checklist above. Below, I provide further context for checkboxes that I didn't check.

General checks

- **Substantial scholarly effort**: following the review guidelines at https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort, here are my observations on this: the code was written by a single author over a limited period of time. There are no external stars/watchers/forks/issue submitters for the project so usage has not yet been demonstrated. Despite this, the quantitative metrics over the code (i.e., number of lines) and my own review of the code show that the work put into the implementation is not trivial. If these characteristics are consistent with a JOSS publication, I can check this box - guidance from @fboehm would be helpful.

- **Reproducibility**: across the entire repository, documentation and paper, there
is a single piece of example code demonstrating the software, however, this
does not work as expected. One or more demonstration notebooks of the utility of the
software on some example use cases would improve the package and the paper.

Functionality

- **Functionality**: The functionality of the software is not clearly described.
Across the repository's README, the documentation, and the paper there isn't a
single place where a clear and well structured summary exists of the software's
functionalities. This should be improved for clarity and usability.

Documentation
- **Example usage**: There is only one piece of example code at
https://phylox.readthedocs.io/en/latest/index.html but it is not working as
expected: https://github.com/RemieJanssen/PhyloX/issues/42.
- **Functionality documentation**: The documentation is not complete. It should
provide a better way to navigate the software modules from the documentation landing page. When reviewing the code
I found many of the classes and functions in different modules were missing
docstrings, e.g., https://github.com/RemieJanssen/PhyloX/blob/main/src/phylox/generators/heath/heath.py - these then don't show up in the documentation.
- **Community guidelines**: The repository provides a CONTRIBUTING.md description at https://github.com/RemieJanssen/PhyloX/blob/main/CONTRIBUTING.md. However, the description of Pull requests is not perfect: it assumes a contributor can
git clone git@github.com:RemieJanssen/PhyloX.git and then push to origin. This will generally not work for outside contributors. The contribution guidelines should therfore mention forking the project to contribute.

Software paper
- **Summary**: the current description is not clear, especially for a non-specialist audience. It could be improved by a clearer description of the basic concepts the package uses (e.g., phylogenetic network vs tree, what is "cherry-picking", etc.). Obviously one can't expect to go into details so this would have to be succinct and to the point. Perhaps approaching this by more clearly stating the kinds of analysis the package enables could also help.

- **Quality of writing**: Currently more than half of the paper is discussing other related tools. This should be rebalanced by focusing on PhyloX and giving a proper description of it. I also found some typos and language issues that I submitted a PR to improve.
