### 2024-03-24 ae20b44
add seed in example and fix height related problems in random network
