### 2024-03-24 20c2aea
fix add_root_if_necessary and simple newick printing
