### 2024-03-24 RemieJanssen/PhyloX#42 @RemieJanssen
It seems to be a problem with deepcopying of networks in combination with caching of some network properties.
