### 2024-03-24 4f87610
slightly shorter test. added example from docs as test
