### 2024-03-26 f90e782
Improvements to JOSS paper (#44)
Fixes typos and some odd language in the text
