+++ joss/paper.md
@@ -18,18 +18,18 @@ bibliography: paper.bib
 
 # Summary
 
-PhyloX is a Python package with tools for generating, manipulating, and analyzing phylogenetic networks. It uses the NetworkX package [@SciPyProceedings_11] for basic graph operations. This has the added benefit that the powerful graph tools from networkX can be used directly on the phylogenetic networks as well. The aim of the package is to be of general use to phylogenetic network researchers, with a current focus on I/O, random generation of networks, cherry-picking methods, rearrangement operations, and the identification of classes and properties of networks.
+PhyloX is a Python package with tools for generating, manipulating, and analyzing phylogenetic networks. It uses the NetworkX package [@SciPyProceedings_11] for basic graph operations. This has the added benefit that the powerful graph tools from NetworkX can be used directly on the phylogenetic networks as well. The aim of the package is to be of general use to phylogenetic network researchers, with a current focus on I/O, random generation of networks, cherry-picking methods, rearrangement operations, and the identification of classes and properties of networks.
 
 # Statement of Need
-Currently, there is no Python package that enables a full workflow for analysing properties and methods of phylogenetic networks. Isolated scripts for this purpose do appear on GitHub or as pseudo-code regularly, most often as part of publications studying one method or one property [@zhang2018bayesian; @pons2019generation; @janssen2020combining; @janssen2020linear; @janssen2021rearranging]. Combining such scripts requires quite some work, for example because the phylogenetic networks themselves are represented by different Python classes with their own methods.
+Currently, there is no Python package that enables a full workflow for analysing properties and methods of phylogenetic networks. Isolated scripts for this purpose do appear on GitHub or as pseudo-code regularly, most often as part of publications studying one method or one property [@zhang2018bayesian; @pons2019generation; @janssen2020combining; @janssen2020linear; @janssen2021rearranging]. Combining such scripts requires substantial work, for example because the phylogenetic networks themselves are represented by different Python classes with their own methods.
 
-This package, PhyloX, aims to bring these scripts together: it standardizes implementations of several basic objects related to pylogenetic networks, such as the networks themselves, the labelling of the nodes, and rearrangement moves. It currently implements a limited but important set of basic functions: I/O for networks (e.g. lists of edges and extended newick format), network generation for test sets, comparing networks resulting from reconstruction methods, and computing several well-used network properties such as the reticulation number, the level, and the number of cherries.
+This package, PhyloX, aims to bring these scripts together: it standardizes implementations of several basic objects related to phylogenetic networks, such as the networks themselves, the labelling of the nodes, and rearrangement moves. It currently implements a limited but important set of basic functions: I/O for networks (e.g. lists of edges and extended newick format), network generation for test sets, comparing networks resulting from reconstruction methods, and computing several well-used network properties such as the reticulation number, the level, and the number of cherries.
 
 # Related packages
 As mentioned above, there are currently no Python packages that enable a complete workflow for phylogenetic networks. There are, however, a few Python packages that seem to fit that bill to a certain extent. We will argue that, despite these packages being available, there is still a need, or at least a great benefit, of using PhyloX.
 
 ## PhyloNetwork
[…]
