### 2024-03-26 957c41d
Fix main example docs (#45)
* add seed in example and fix height related problems in random network

* fix add_root_if_necessary and simple newick printing

* fix issue with main example in docs: use leaves from copy instead of original in is_orchard

* slightly shorter test. added example from docs as test

* CHANGELOG
