+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Types of changes:
 
 ## [1.0.4] - (Unreleased)
 
+- Fixed main example from the documentation. (#42)
+
 ## [1.0.3] - (2024-01-20)
 
 ### Changed
