### 2024-03-26 RemieJanssen/PhyloX#42 @RemieJanssen
@abhishektiwari @bgyori I fixed the example and merged it into the main branch already. Feel free to test it again. If you would rather do that with a new version via pip, I can make a new release. Though, I'd prefer to bundle it with some more fixes, if you find any more bugs.

I'll close this issue once any of you can confirm the example works.
