### 2024-04-06 RemieJanssen/PhyloX#46 @abhishektiwari
Update _autosummary RST to include missing health generator module
**Describe the bug**
Read the doc currently missing [heath](https://github.com/RemieJanssen/PhyloX/tree/main/src/phylox/generators/heath) generator module.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6427)
