### 2024-04-06 #6427 @abhishektiwari
@RemieJanssen

Please see more feedback

## Quality of writing
Some parts need rewriting like related packages (State of the field). It can be more crisp, bit shorter (as suggested by other reviewer), and objective. I will avoid using weasel phrases like _fit that bill to a certain extent_, and be more specific. I felt comparison on the documentation page did a better job by calling out the specific differences between PhyloX and alternatives.

> Unlike other packages like [Biopython](https://biopython.org/) and [DendroPy](https://dendropy.org/), PhyloX is designed to work with phylogenetic networks, rather than phylogenetic trees.

I also suggest highlighting some of the key features/functionalities of the PhyloX software in paper. Again, I think documentation page does a better job describing key features of PhyloX.

## Example usage
Thanks for fixing the example. I would love to see one or two Jupyter notebook based example use cases. But this is not a blocker from my side.

## API documentation
Module functions for the health generator are missing please [fix RST tree](https://github.com/RemieJanssen/PhyloX/issues/46).
