### 2024-04-15 #6427 @fboehm
> @RemieJanssen - do you have questions about next steps? Thanks again!

Sorry that I missed your earlier messsge. Yes, compared to some review processes, that of Joss may feel different. It is essentially an iterative process between you and the reviewers. Once they are able to check all boxes, their roles are complete. As author, you just need to respond to their suggestions. If you disagree with a suggestion, please discuss it in this thread.
