### 2024-04-20 #6427 @bgyori
I'm just checking back again, @RemieJanssen have you had a chance to look into the paper improvements / documentation improvements that @abhishektiwari and I suggested? I don't see any recent changes since my PR was merged with some typo fixes, though I might have missed something.
