### 2024-04-21 #6427 @RemieJanssen
> > @RemieJanssen - do you have questions about next steps? Thanks again!
>
> Sorry that I missed your earlier messsge. Yes, compared to some review processes, that of Joss may feel different. It is essentially an iterative process between you and the reviewers. Once they are able to check all boxes, their roles are complete. As author, you just need to respond to their suggestions. If you disagree with a suggestion, please discuss it in this thread.

Ah! That does answer my question, thank you :) I finally have some time again, so I'll try to get as far as possible responding to all the comments this week.
