+++ src/phylox/generators/heath/heath_cli.py
@@ -9,26 +9,27 @@ import numpy as np
 import scipy.stats
 
 # from AddEdgesToTree import *
-from phylox.generators.heath.heath import RestrictToLeafSet, generate_heath_network
+from phylox.generators.heath.heath import restrict_network_to_leaf_set, generate_heath_network
 
 
 def parse_args():
     parser = argparse.ArgumentParser(
         description="""
         Runs a speciation-extinction-HGT-hybridization model for the given time (--time) or until a certain number of extant taxa (--taxa) is reached. If all lineages go extinct before the given time is reached, another attempt is made.
-        
-        Each extant taxon has its own speciation, HGT, and extinction rates (1/mean_time_until_next_event). 
-        
-        The hybridization rate of a pair of species is a function of the weighed distance between these species (sum of all up-down distances, weighed by their probability). 
-        
-        There are prior speciation/extinction/HGT rate distributions: gamma distributions with a given mean and a shape parameter for speciation (--speciation_parameters) and extinction (--extinction_parameters), HGT is turned off by default. 
-        
-        If a HGT event happens for a given taxon, another taxon (including itself) is chosen uniformly at random to donate genetic material (uniformly distributed contribution in [0,max_hgt] where max_hgt is determined by the --hgt_inheritance parameter). After speciation or hybridization, each rate of the new lineages is set by multiplying the (weighted mean) rate of the parent lineage(s) by a gamma-distributed factor with mean 1 and a shape parameter (--update-shape-parameter), and then accepting this rate with a probability proportional to the prior distribution for this rate. This gives an ultrametric network on the extant species.
-        
+
+        Each extant taxon has its own speciation, HGT, and extinction rates (1/mean_time_until_next_event).
+
+        The hybridization rate of a pair of species is a function of the weighed distance between these species (sum of all up-down distances, weighed by their probability).
+
+        There are prior speciation/extinction/HGT rate distributions: gamma distributions with a given mean and a shape parameter for speciation (--speciation_parameters) and extinction (--extinction_parameters), HGT is turned off by default.
+
+        If an HGT event happens for a given taxon, another taxon (including itself) is chosen uniformly at random to donate genetic material (uniformly distributed contribution in [0,max_hgt] where max_hgt is determined by the --hgt_inheritance parameter). After speciation or hybridization, each rate of the new lineages is set by multiplying the (weighted mean) rate of the parent lineage(s) by a gamma-distributed factor with mean 1 and a shape parameter (--update-shape-parameter), and then accepting this rate with a probability proportional to the prior distribution for this rate. This gives an ultrametric network on the extant species.
+
         Optional arguments:
-            -ti or --time followed by the total length (float) for the network. 
+            -ti or --time followed by the total length (float) for the network.
             -ta or --taxa followed by the number of taxa at which the simulation stops. If all lineages go extinct before the given number of taxa is reached, another attempt is made.
-            -ce or --count_extinct to also count the extinct taxa as part of the taxa limit.-oe or --only_extant to return the network restricted to the extant leaves
+            -ce or --count_extinct to also count the extinct taxa as part of the taxa limit.
+            -oe or --only_extant to return the network restricted to the extant leaves
         Rates:
             -sp or --speciation_parameters followed by a mean (float) and a shape parameter (float) for the gamma distribution of the speciation rate.
             -ext or --extinction_parameters followed by a mean (float) and a shape parameter (float) for the gamma distribution of the extinction rate.
@@ -41,16 +42,17 @@ def parse_args():
         right bound: r
[…]
