+++ src/phylox/generators/heath/heath.py
@@ -31,26 +31,38 @@ import scipy.stats
 #  for an extinction event, the distances simply increase by two times the time to speciation, and the distances to the extinct species are removed.
 #  for a HGT event, only the rates for the receiving taxon have to be updated. They can be computed from the distances to the receiving and the donating easily
 #  for a hyb event, all paths go via exactly one of the parent species, so we can use those distances again.
-def UpdateHybridizationRates():
+def update_hybridization_rates():
     return False
     # This is done within the CalculateNetwork function
 
 
-def GammaDistributionPDF(value, mean, shape):
+def gamma_distribution_pdf(value, mean, shape):
+    """
+    A reparameterization of the Gamma Distribution from scipy stats,
+    with mean parameter instead of a shape parameter.
+
+    """
     scale = mean / shape
     return scipy.stats.gamma.pdf(value, shape, 0, scale)
 
 
-def CalculateNewRate(parent_rate, prior_mean, prior_shape, update_shape):
+def update_rate(parent_rate, prior_mean, prior_shape, update_shape):
+    """
+    Generates the new value for a rate used in the Heath generator based on
+     - the current value,
+     - the multiplicative update factor `random.gammavariate(<update_shape>, 1 / <update_shape>)`
+     - the prior Gamma Distribution for the rate type `GammaDistributionPDF(rate_value, prior_mean, prior_shape)`
+    """
+
     proposed_rate = parent_rate * random.gammavariate(update_shape, 1 / update_shape)
-    if random.random() < GammaDistributionPDF(
+    if random.random() < gamma_distribution_pdf(
         proposed_rate, prior_mean, prior_shape
-    ) / GammaDistributionPDF(parent_rate, prior_mean, prior_shape):
+    ) / gamma_distribution_pdf(parent_rate, prior_mean, prior_shape):
         return proposed_rate
     return parent_rate
 
 
-def CalculateAllNewRates(
+def update_all_rates(
     parent_rates,
     update_shape,
     speciation_rate_mean,
@@ -62,30 +74,54 @@ def CalculateAllNewRates(
     hgt_rate_mean,
     hgt_rate_shape,
 ):
-    sp_rate = CalculateNewRate(
+    """
+    Updates all rates for the Heath generator using the `update_rate` function.
+    """
+    sp_rate = update_rate(
         parent_rates[0], speciation_rate_mean, speciation_rate_shape, update_shape
     )
     ext_rate = 0
     if ext_used:
-        ext_rate = CalculateNewRate(
+        ext_rate = update_rate(
             parent_rates[1], extinction_rate_mean, extinction_rate_shape, update_shape
         )
     hgt_rate = 0
     if hgt_used:
-        hgt_rate = CalculateNewRate(
+        hgt_rate = update_rate(
             parent_rates[2], hgt_rate_mean, hgt_rate_shape, update_shape
         )
     return (sp_rate, ext_rate, hgt_rate)
 
 
-def DistanceToRate(
+def graph_distance_to_hybridization_rate(
     distance,
     hybridization_left_bound,
     hybridization_right_bound,
     hybridization_left_rate,
     hybridization_right_rate,
 ):
-    # TODO Update this rate function, make it more of a gradual switch?
+    """
+    Returns a hybdridization rate for a given distance.
+    The distance represents the evolutionary distance between two extant species.
+    The function is used in the Heath generator to determine the hybdridization rate between two current taxa,
+    where the distance is a weighthed sum of all up-down distances between the two taxa in the current network.
+    The dependency on the distance takes the following shape:
+        left bound:  l
+        right bound: r
+        left rate:   rl
+        right rate:  rr.
+
+           |
+        lr +---
+           |   \\
+           |    \\
+           |     \\
+        rr +      -----
+           |
+         0 +---+----+-----
[…]
