+++ src/phylox/rearrangement/movetype.py
@@ -1,6 +1,5 @@
 from enum import Enum
 
-
 class MoveType(str, Enum):
     NONE = "NONE"
     TAIL = "TAIL"
