+++ docs/source/_autosummary/phylox.generators.heath.heath.RestrictToLeafSet.rst
@@ -1,0 +1,6 @@
+phylox.generators.heath.heath.RestrictToLeafSet
+===============================================
+
+.. currentmodule:: phylox.generators.heath.heath
+
+.. autofunction:: RestrictToLeafSet

[…]
