+++ src/phylox/constants.py
@@ -6,5 +6,8 @@ This module contains some constants used throughout the package.
 LABEL_ATTR = "label"
 #: The edge attribute used to store the length of an edge.
 LENGTH_ATTR = "length"
+#: The edge attribute used to store the inheritance probability along an edge.
+PROBABILITY_ATTR = "probability"
+
 #: Prefix for the node name of a reticulation node when parsing a phylogenetic network from newick
 RETIC_PREFIX = "__#R"
