### 2024-04-21 RemieJanssen/PhyloX#46 @RemieJanssen
I remembered that I intentionally left it out of the documentation when I tried to add it, because I had not really reworked it from the source to the structure of this package. I have done the minimal to add it in a more meaningful way, but I think it can use some more work.

https://github.com/RemieJanssen/PhyloX/pull/48
