+++ docs/source/_autosummary/phylox.generators.heath.heath.restrict_network_to_leaf_set.rst
@@ -1,0 +1,6 @@
+phylox.generators.heath.heath.restrict\_network\_to\_leaf\_set
+==============================================================
+
+.. currentmodule:: phylox.generators.heath.heath
+
+.. autofunction:: restrict_network_to_leaf_set
