+++ docs/source/_autosummary/phylox.generators.heath.heath.GammaDistributionPDF.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.GammaDistributionPDF
-==================================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: GammaDistributionPDF
