+++ docs/source/_autosummary/phylox.generators.heath.heath.DistanceToRate.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.DistanceToRate
-============================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: DistanceToRate
