+++ docs/source/_autosummary/phylox.constants.PROBABILITY_ATTR.rst
@@ -1,0 +1,6 @@
+phylox.constants.PROBABILITY\_ATTR
+==================================
+
+.. currentmodule:: phylox.constants
+
+.. autodata:: PROBABILITY_ATTR
