+++ docs/source/_autosummary/phylox.generators.heath.heath.SuppressDegree2.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.SuppressDegree2
-=============================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: SuppressDegree2
