+++ docs/source/_autosummary/phylox.generators.heath.heath.CalculateNewRate.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.CalculateNewRate
-==============================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: CalculateNewRate
