+++ docs/source/_autosummary/phylox.generators.heath.heath.UpdateHybridizationRates.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.UpdateHybridizationRates
-======================================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: UpdateHybridizationRates
