+++ docs/source/_autosummary/phylox.generators.heath.heath.CalculateAllNewRates.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.CalculateAllNewRates
-==================================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: CalculateAllNewRates
