+++ docs/source/_autosummary/phylox.generators.heath.heath.RestrictToLeafSet.rst
@@ -1,6 +1,0 @@
-phylox.generators.heath.heath.RestrictToLeafSet
-===============================================
-
-.. currentmodule:: phylox.generators.heath.heath
-
-.. autofunction:: RestrictToLeafSet
