+++ src/phylox/generators/heath/heath.py
@@ -205,8 +205,8 @@ def generate_heath_network(
     speciation_rate_shape=2.0,
     ext_used=True,
     count_extinct=False,
-    extinction_rate_mean=2.0,
-    extinction_rate_shape=2.0,
+    extinction_rate_mean=1.0,
+    extinction_rate_shape=1.0,
     hgt_used=False,
     hgt_rate_mean=None,
     hgt_rate_shape=None,
