### 2024-04-21 RemieJanssen/PhyloX#49 @RemieJanssen
Improve JOSS paper
Reviewers have requests for improvements in the paper: https://github.com/openjournals/joss-reviews/issues/6427

**Describe the solution you'd like**
 - More crisp writing for related packages [abishektiwari](https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-2041108193)
 - Summary of features of the package (for non-specialists), i.e., Key features/highlights of the package [abishektiwari](https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-2041108193) [bgyori](https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-2016594428)
 - More examples: [bgyori](https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-2016594428)
 - Stating the kinds of analysis the package enables. [bgyori](https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-2016594428)
