### 2024-04-21 RemieJanssen/PhyloX#50 @RemieJanssen
Fix Contributing Guidelines
**Describe the bug**
Community guidelines are not correct: for contributing, you cannot just pull and create a merge request. People need to fork first. Fix that in the description. See: https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-2016594428
