### 2024-04-22 #6427 @fboehm
> > I have now completed my initial review, see checklist above. Below, I provide further context for checkboxes that I didn't check.
> > General checks
> > ```
> > * **Substantial scholarly effort**: following the review guidelines at https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort, here are my observations on this: the code was written by a single author over a limited period of time. There are no external stars/watchers/forks/issue submitters for the project so usage has not yet been demonstrated. Despite this, the quantitative metrics over the code (i.e., number of lines) and my own review of the code show that the work put into the implementation is not trivial. If these characteristics are consistent with a JOSS publication, I can check this box - guidance from @fboehm would be helpful.
> > ```
>
> I don't think I can really argue with you here, most of the package was written by me, and compiled over a short period of time. However, it's not really true that it was written over a limited period. The package is a compilation of most code was written during the 4 years of my PhD and the few years after (and not even all by me, but also by co-authors). This package is an attempt to not let all that effort go to waste because of our unFAIR code practices.

@RemieJanssen @bgyori - can you elaborate on this point? Why does it appear to @bgyori that things were written over a short period while @RemieJanssen says otherwise? Is @bgyori relying on git commit times, while @RemieJanssen and co-workers perhaps didn't use git while writing?
