### 2024-04-23 RemieJanssen/PhyloX#51 @RemieJanssen
Missing docstrings in `phylox.generators.lgt.base`
Missing docstrings in `phylox.generators.lgt.base` https://github.com/openjournals/joss-reviews/issues/6427

**Describe the solution you'd like**
Write docstrings for the helper functions in that module
