+++ joss/paper.bib
@@ -91,3 +91,129 @@
   year={2020},
   organization={Springer}
 }
+
+@article{kong2022classes,
+  title={Classes of explicit phylogenetic networks and their biological and mathematical significance},
+  author={Kong, Sungsik and Pons, Joan Carles and Kubatko, Laura and Wicke, Kristina},
+  journal={Journal of Mathematical Biology},
+  volume={84},
+  number={6},
+  pages={47},
+  year={2022},
+  publisher={Springer}
+}
+
+@ARTICLE{4407681,
+  author={Cardona, Gabriel and Rossello, Francesc and Valiente, Gabriel},
+  journal={IEEE/ACM Transactions on Computational Biology and Bioinformatics},
+  title={Comparison of Tree-Child Phylogenetic Networks},
+  year={2009},
+  volume={6},
+  number={4},
+  pages={552-569},
+  keywords={Phylogeny;History;Genetic mutations;Reconstruction algorithms;Polynomials;Computer networks;Packaging;Java;Organisms;Databases;Phylogenetic network;tree-child phylogenetic network;phylogenetic tree;partition distance;network alignment.;Biology and genetics;Life and Medical Sciences;Computer Applications;Graph algorithms;Graph Theory;Discrete Mathematics;Mathematics of Computing;Trees;Graph Theory;Discrete Mathematics;Mathematics of Computing},
+  doi={10.1109/TCBB.2007.70270}}
+
+
+@book{huson2010phylogenetic,
+  title={Phylogenetic networks: concepts, algorithms and applications},
+  author={Huson, Daniel H and Rupp, Regula and Scornavacca, Celine},
+  year={2010},
+  publisher={Cambridge University Press}
+}
+
+@article{ERDOS201933,
+title = {A class of phylogenetic networks reconstructable from ancestral profiles},
+journal = {Mathematical Biosciences},
+volume = {313},
+pages = {33-40},
+year = {2019},
+issn = {0025-5564},
+doi = {https://doi.org/10.1016/j.mbs.2019.04.009},
+url = {https://www.sciencedirect.com/science/article/pii/S0025556419300239},
+author = {Péter L. Erdős and Charles Semple and Mike Steel},
+keywords = {Tree-child networks, Orchard networks, Accumulation phylogenies, Ancestral profiles, Path-tuples},
[…]
