+++ joss/paper.md
@@ -8,7 +8,7 @@ tags:
 authors:
   - name: Remie Janssen
     orcid: 0000-0002-5192-1470
-    affiliation: 1 
+    affiliation: 1
 affiliations:
  - name: National Institute for Public Health and the Environment, Bioinformatics and Computing group, Bilthoven, The Netherlands
    index: 1
@@ -20,10 +20,46 @@ bibliography: paper.bib
 
 PhyloX is a Python package with tools for generating, manipulating, and analyzing phylogenetic networks. It uses the NetworkX package [@SciPyProceedings_11] for basic graph operations. This has the added benefit that the powerful graph tools from NetworkX can be used directly on the phylogenetic networks as well. The aim of the package is to be of general use to phylogenetic network researchers, with a current focus on I/O, random generation of networks, cherry-picking methods, rearrangement operations, and the identification of classes and properties of networks.
 
+# Phylogenetic networks
+In the study of the evolutionary history of biological species and languages, it is common to represent putative histories using graphs. Traditionally, at least in biology, these graphs were most often trees, such as the well known tree drawn by Charles Darwin in one of his notebooks. A tree like this called a phylogenetic tree. In some cases, the evolutionary history includes complex processes like horizontal gene transfer and hybridization. These processes cause a *reticulate* (i.e. network-like) structure in the evolutionary history, which requires phylogenetic networks to be used for representing the evolutionary histories.
+
+A *directed phylogenetic network* (e.g., [@huson2010phylogenetic]) is a directed acyclic graph with four types of nodes:
+ - a *root*: an in-degree 0, out-degree 1 node,
+ - a labelled set of *leaves*: in-degree 1, out-degree 0 nodes,
+ - a set of *reticulation nodes*: in-degree $>1$, out-degree 1 nodes,
+ - a set of *tree nodes*: in-degree 1, out-degree $>1$ nodes.
+
+A network is *binary* if each reticulation node has in-degree 2, and each tree node has in-degree 2. An *undirected phylogenetic network* is the underlying undirected graph of a directed phylogenetic network, retaining the labelling of the leaf nodes.
+
+## Network properties
+When analysing or comparing phylogenetic networks or phylogenetic network methods, it is often useful to extract some (numerical) parameters from the networks. Some of the most used properties are the *reticulation number* (the number of reticulation nodes in a binary network), the number of *blobs* (biconnected components of the network), and the *level* (the maximum reticulation numberamong all blobs of the network). Of course, there are more, like the recently introduced $B_2$-balance index of the network [@franccois2021revisiting].
+
+## Classes of networks
+In research on phylogenetic networks, it is common to restrict attention to some well-known classes of phylogenetic networks. These classes put additional restrictions on the definition of a network, for the benefit of computational efficiency, to model certain biological restraints, or for both.
+
+Kong et al. [@kong2022classes] gives a good overview of most well-known classes of directed phylogenetic networks, and their biological interpretation. For example, tree-child networks are networks in where each ancestral species has a descendant among the extant taxa (the leaves) through only mutation in the network [@4407681]. Mathematically, they are characterized as networks in which each non-leaf node has at least one child that is not a reticulation node.
+
+## Cherry-picking
+A basic structure in any network or tree is the *cherry*, a pair of leaves with a common parent. A modified version often found in phylogenetic networks is the *reticulated cherry*, an ordered pair of leaves $(x, y)$ which are related through the three edges $(p_x, x)$, $(p_y, y)$, and $(p_y, p_x)$.
+
[…]
