+++ joss/paper.md
@@ -46,12 +46,40 @@ A common modification to a phylogenetic network is to *pick* or ( or *reduce*) a
 
 These modifications are used in computational tools, for example to reconstruct networks from ancestral profiles [@ERDOS201933; @bai2021defining; @cardona2024comparison], to check whether one tree-child network is contained in another [@janssen2021cherry], or to combine multiple trees into one network [@linz2019attaching; @van2022practical]. Their versatile use also lead to the introduction of the class of *orchard networks* [@janssen2021cherry; @ERDOS201933]. This class contains all networks that can be reduced to a single leaf using cherry-picking operations. Networks from this class can be interpreted as trees with horizontal gene transfer arcs [@van2022orchard].
 
+## Rearranging networks
+For phylogenetic inference problems, it is common to use heuristics that search through a space of networks. Such a space of networks takes the shape of a graph, whose objects are all networks with a common set of leaf labels (the sampled taxa), and sometimes also a set number of reticulations. The edges of the graph correspond to small changes made to a network: there is an edge between two networks if one can make a modification to one of the networks to arrive at the second network.
+
+The modifications that are allowed are well-defined as types of *rearrangement* opterations. These operations can be *horizontal*, keeping the reticulation number the same, or *vertical*, changing the reticulation number. Most horizontal operations are a variation on or restriction of the *rooted subtree prune and regraft* (*rSPR*) operation [@bordewich2017lost; @gambette2017rearrangement].
+
+The names for vertical moves have not been standardized, but they generally do the same. A vertical move that removes a reticulation removes an incoming edge of a reticulation node, and then suppresses the resulting degree 2 nodes. A vertical move that adds a reticulation does the reverse: it *subdivides* two edges of the network, and adds a new edge between the two new degree 2 nodes (e.g. [@bordewich2017lost]).
+
+As mentioned, rearrangement moves can be used to traverse a space of networks. This is used, for example, to sample posterior distributions in Bayesian analyses [@wen2016bayesian, @zhang2018bayesian] and to find networks that score high on a maximum likelihood criterium [@wen2016reticulate; @yu2014maximum].
+
 ## Generating networks
 To test phylogenetic network methods, one either needs to source or create a test set of networks. Creating them is often the simpler option, so methods to randomly generate phylogenetic networks are ready at hand. Moreover, these methods are often based on evolutionary models that are defined on a high level, i.e. with explicit events for  with processes such as speciation, extinction, and hybridization.
 
 The paper [@janssen2021comparing] contains a comparison of several 'generators', including several previously existing ones (e.g., [@pons2019generation] and [@zhang2018bayesian]) and a new extention of a tree generator to networks.
 
+## Representing networks
+There are some common representations of phylogenetic networks. As they are graphs, there is the obvious option to represent them as a list of edges. Another commonly used representation is the extended Newick format [@cardona2008extended]. The extended Newick notation has a further extension (Rich Newick format) that adds numerical parameters to the edges of the network, such as the branch length, and the inheritance probability (for incoming edges of a reticulation node) [@riceRichNewick; @wen2018inferring].
+
 # PhyloX Functionality
+
+PhyloX is equipped to handle all the aspects pf phylogenetic networks mentioned in the previous section.
+
+## I/O
+- similar to networkx, but with labels
+- from/to newick
+- images (networkx)
+
+## Generating networks
+ - random seeds
+ - many of the known phylogenetic network generators
+   - MCMC with classes [@klawitter2020spaces]
+
+
+## Comparing networks
[…]
