+++ joss/paper.bib
@@ -217,3 +217,102 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
   year={2021},
   publisher={World Scientific}
 }
+
+@article{cardona2008extended,
+  title={Extended Newick: it is time for a standard representation of phylogenetic networks},
+  author={Cardona, Gabriel and Rossell{\'o}, Francesc and Valiente, Gabriel},
+  journal={BMC bioinformatics},
+  volume={9},
+  pages={1--8},
+  year={2008},
+  publisher={Springer}
+}
+
+@misc{riceRichNewick,
+	author = {},
+	title = {{R}ich {N}ewick {F}ormat - {P}hylonet - {R}ice {U}niversity {C}ampus {W}iki --- wiki.rice.edu},
+	howpublished = {\url{https://wiki.rice.edu/confluence/display/PHYLONET/Rich+Newick+Format}},
+	year = {},
+	note = {[Accessed 25-04-2024]},
+}
+
+@article{wen2018inferring,
+  title={Inferring phylogenetic networks using PhyloNet},
+  author={Wen, Dingqiao and Yu, Yun and Zhu, Jiafan and Nakhleh, Luay},
+  journal={Systematic biology},
+  volume={67},
+  number={4},
+  pages={735--740},
+  year={2018},
+  publisher={Oxford University Press}
+}
+
+@article{wen2016reticulate,
+  title={Reticulate evolutionary history and extensive introgression in mosquito species revealed by phylogenetic network analysis},
+  author={Wen, Dingqiao and Yu, Yun and Hahn, Matthew W and Nakhleh, Luay},
+  journal={Molecular ecology},
+  volume={25},
+  number={11},
+  pages={2361--2372},
+  year={2016},
+  publisher={Wiley Online Library}
+}
+
+@article{yu2014maximum,
+  title={Maximum likelihood inference of reticulate evolutionary histories},
+  author={Yu, Yun and Dong, Jianrong and Liu, Kevin J and Nakhleh, Luay},
+  journal={Proceedings of the National Academy of Sciences},
+  volume={111},
+  number={46},
+  pages={16448--16453},
+  year={2014},
+  publisher={National Acad Sciences}
+}
+
+@article{wen2016bayesian,
+  title={Bayesian inference of reticulate phylogenies under the multispecies network coalescent},
+  author={Wen, Dingqiao and Yu, Yun and Nakhleh, Luay},
+  journal={PLoS genetics},
+  volume={12},
+  number={5},
+  pages={e1006006},
+  year={2016},
+  publisher={Public Library of Science San Francisco, CA USA}
+}
+
+@article{bordewich2017lost,
+  title={Lost in space? Generalising subtree prune and regraft to spaces of phylogenetic networks},
+  author={Bordewich, Magnus and Linz, Simone and Semple, Charles},
+  journal={Journal of theoretical biology},
+  volume={423},
+  pages={1--12},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@phdthesis{klawitter2020spaces,
+  title={Spaces of phylogenetic networks},
+  author={Klawitter, Jonathan},
+  year={2020},
+  school={University of Auckland}
+}
+
+@phdthesis{thesis_janssen,
+    author = "Janssen, Remie",
+    school = "Delft University of Technology",
+    title = {Rearranging phylogenetic networks},
+    year = 2021,
+    DOI = "https://doi.org/10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57",
+    ISBN = "978-94-6332-758-9"
+}
+
+@article{gambette2017rearrangement,
+  title={Rearrangement moves on rooted phylogenetic networks},
[…]
