+++ joss/paper.bib
@@ -316,3 +316,26 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
   year={2017},
   publisher={Public Library of Science San Francisco, CA USA}
 }
+
+@misc{https://doi.org/10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1,
+  doi = {10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1},
+  url = {},
+  author = {Julien, Esther and Bernardini, Giulia and Stougie, Leen and van Iersel, Leo},
+  keywords = {Phylogenetics, Hybridization, Cherry Picking, Machine Learning, Heuristic},
+  title = {Source code for the paper &#34;Constructing Phylogenetic Networks via Cherry Picking and Machine Learning&#34;},
+  publisher = {4TU.ResearchData},
+  year = {2023},
+  copyright = {MIT},
+}
+
+
+@article{bernardini2023constructing,
+  title={Constructing phylogenetic networks via cherry picking and machine learning},
+  author={Bernardini, Giulia and van Iersel, Leo and Julien, Esther and Stougie, Leen},
+  journal={Algorithms for Molecular Biology},
+  volume={18},
+  number={1},
+  pages={13},
+  year={2023},
+  publisher={Springer}
+}
