+++ joss/paper.bib
@@ -229,11 +229,12 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
 }
 
 @misc{riceRichNewick,
-	author = {},
-	title = {{R}ich {N}ewick {F}ormat - {P}hylonet - {R}ice {U}niversity {C}ampus {W}iki --- wiki.rice.edu},
-	howpublished = {\url{https://wiki.rice.edu/confluence/display/PHYLONET/Rich+Newick+Format}},
-	year = {},
-	note = {[Accessed 25-04-2024]},
+  title={Rich newick format},
+  url={https://wiki.rice.edu/confluence/display/PHYLONET/Rich+Newick+Format},
+  journal={Rich Newick Format - Phylonet - Rice University Campus Wiki},
+  author={Barnett, Robert},
+  year={2012},
+  month={Feb}
 }
 
 @article{wen2018inferring,
