+++ joss/paper.bib
@@ -340,3 +340,28 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
   year={2023},
   publisher={Springer}
 }
+
+@article{ERDOS2021205,
+title = {Rooted NNI moves and distance-1 tail moves on tree-based phylogenetic networks},
+journal = {Discrete Applied Mathematics},
+volume = {294},
+pages = {205-213},
+year = {2021},
+issn = {0166-218X},
+doi = {https://doi.org/10.1016/j.dam.2021.02.016},
+url = {https://www.sciencedirect.com/science/article/pii/S0166218X21000597},
+author = {Péter L. Erdős and Andrew Francis and Tamás Róbert Mezei},
+keywords = {Nearest-neighbour interchange, Rooted NNI moves, Distance-1 tail moves, Rooted phylogenetic network, Directed acyclic graph, Tree-based network},
+abstract = {We show that any two tree-based rooted binary phylogenetic networks (with the same number of reticulation nodes and identical degrees at the root) are connected by at most (4k+8)⋅|X|+34k+|X|⋅⌈log2|X|⌉ rooted nearest-neighbour interchange (rNNI) moves. As a corollary of the proof we obtain that the diameter is at most a 2k additive term larger for distance-1 tail moves.}
+}
+
+@article{van2022orchard,
+  title={Orchard networks are trees with additional horizontal arcs},
+  author={van Iersel, Leo and Janssen, Remie and Jones, Mark and Murakami, Yukihiro},
+  journal={Bulletin of Mathematical Biology},
+  volume={84},
+  number={8},
+  pages={76},
+  year={2022},
+  publisher={Springer}
+}
