### 2024-04-25 54621fa
improve joss paper (#52)
* add intro phylogenetic networks

* changelog

* add rearrangement, start phylox functionality, and edit statement of need

* intro for functionality section

* add I/O section

* generators described

* fix typo and correct indentation in example

* clarify example

---------
