### 2024-04-25 #6427 @RemieJanssen
I tried to re-balance the paper as @bgyori suggested, in this PR: https://github.com/RemieJanssen/PhyloX/pull/52
Please let me know if this is the direction you were thinking of. If it is, I will go through the paper again to clean up the text (language and structure), but I'd first like to know if these additions are wanted/necessary/appreciated.
