### 2024-04-25 RemieJanssen/PhyloX#53 @RemieJanssen
Missing docstring for `from_cherry_picking_sequence`
**Describe the bug**
Missing docstring for `from_cherry_picking_sequence` in the `CherryPickingMixin`
