### 2024-04-25 #6427 @RemieJanssen
@bgyori
Your comment on the Community guidelines is handled in https://github.com/RemieJanssen/PhyloX/issues/50

About the Functionality documentation. Did you have any problems navigating the docs besides the issue of missing docstrings?
The missing docstrings will be fixed in
 - https://github.com/RemieJanssen/PhyloX/issues/46
 - https://github.com/RemieJanssen/PhyloX/issues/51
 - https://github.com/RemieJanssen/PhyloX/issues/53

I think that is all of them (except some helper functions for API methods), please let me know if you find any more
