### 2024-04-27 9d2251c
more flexible newick parser
