+++ tests/test_parser.py
@@ -29,6 +29,21 @@ class TestExtendedNewickToDiNetwork(unittest.TestCase):
         self.assertEqual(network[parent_a][node_a]["length"], 1.0)
 
     def test_small_tree_with_more_attrs(self):
+        newick = "(a:::3.0,b:1.1,(c,d:1.3)::1.4:);"
+        network = extended_newick_to_dinetwork(newick)
+        print(network.edges)
+        network2 = DiNetwork(
+            edges=[(1, 2), (1, 3), (1, 4), (4, 5), (4, 6)],
+            labels=[(2, "a"), (3, "b"), (5, "c"),(6, "d")],
+        )
+        self.assertTrue(is_isomorphic(network, network2))
+        node_a = network.label_to_node_dict["a"]
+        parent_a = network.parent(node_a)
+        self.assertEqual(network[parent_a][node_a].get("length"), None)
+        self.assertEqual(network[parent_a][node_a].get("bootstrap"), None)
+        self.assertEqual(network[parent_a][node_a]["probability"], 3.0)
+
+    def test_small_tree_with_partial_attrs(self):
         newick = "(a:1.0:2.0:3.0,b:1.1,(c:1.2,d:1.3):1.4);"
         network = extended_newick_to_dinetwork(newick)
         network2 = DiNetwork(
@@ -41,6 +56,7 @@ class TestExtendedNewickToDiNetwork(unittest.TestCase):
         self.assertEqual(network[parent_a][node_a]["length"], 1.0)
         self.assertEqual(network[parent_a][node_a]["bootstrap"], 2.0)
         self.assertEqual(network[parent_a][node_a]["probability"], 3.0)
+
 
     def test_network(self):
         newick = "(a,(b)#R1,(#H1,c));"
