### 2024-04-27 81d82d5
more flexible newick parser (#55)
* more flexible newick parser
