+++ src/phylox/parser.py
@@ -227,12 +227,15 @@ def _label_and_attrs_to_dict(label_and_attrs):
                 "length": float(attrs[0]),
             }
         elif len(attrs) == 3:
-            attrs_dict = {
-                "label": label,
-                "length": float(attrs[0]),
-                "bootstrap": float(attrs[1]),
-                "probability": float(attrs[2]),
-            }
+
+            attrs_dict = {"label": label}
+            if attrs[0]:
+                attrs_dict["length"] = float(attrs[0])
+            if attrs[1]:
+                attrs_dict["bootstrap"] = float(attrs[1])
+            if attrs[2]:
+                attrs_dict["probability"] = float(attrs[2])
+
     if "#" in attrs_dict["label"]:
         label, retic_id = attrs_dict["label"].split("#")
         attrs_dict["label"] = label
