### 2024-05-13 31fb1d2
fix CONTRIBUTING.md (#54)
* fix CONTRIBUTING.md

* refer to github docs
