### 2024-05-13 67a5eb9
Add heath generator (#48)
* add heath generator as original code

* add seed for heath, and output as phylox network

* add cli for heath generator
