+++ src/phylox/cherrypicking/base.py
@@ -533,7 +533,7 @@ class CherryPickingMixin:
     @classmethod
     def from_cherry_picking_sequence(cls, sequence, heights=None, label_leaves=True):
         network = cls()
-        heights = heights or [[h, h] for h in range(1,len(sequence)+1)]
+        heights = heights or [[h, h] for h in range(1, len(sequence) + 1)]
         for pair, height in zip(reversed(sequence), reversed(heights)):
             add_pair(
                 network, *pair, height=height, inplace=True, nodes_by_label=label_leaves
