+++ docs/source/_autosummary/phylox.generators.heath.heath.graph_distance_to_hybridization_rate.rst
@@ -1,0 +1,6 @@
+phylox.generators.heath.heath.graph\_distance\_to\_hybridization\_rate
+======================================================================
+
+.. currentmodule:: phylox.generators.heath.heath
+
+.. autofunction:: graph_distance_to_hybridization_rate
