+++ docs/source/_autosummary/phylox.generators.heath.heath.suppress_degree_two_nodes.rst
@@ -1,0 +1,6 @@
+phylox.generators.heath.heath.suppress\_degree\_two\_nodes
+==========================================================
+
+.. currentmodule:: phylox.generators.heath.heath
+
+.. autofunction:: suppress_degree_two_nodes
