+++ docs/source/_autosummary/phylox.constants.rst
@@ -12,6 +12,7 @@ phylox.constants
    
       LABEL_ATTR
       LENGTH_ATTR
+      PROBABILITY_ATTR
       RETIC_PREFIX
    
    
