+++ docs/source/_autosummary/phylox.generators.heath.heath.gamma_distribution_pdf.rst
@@ -1,0 +1,6 @@
+phylox.generators.heath.heath.gamma\_distribution\_pdf
+======================================================
+
+.. currentmodule:: phylox.generators.heath.heath
+
+.. autofunction:: gamma_distribution_pdf
