+++ docs/source/_autosummary/phylox.generators.heath.heath.rst
@@ -1,0 +1,33 @@
+phylox.generators.heath.heath
+=============================
+
+.. automodule:: phylox.generators.heath.heath
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      gamma_distribution_pdf
+      graph_distance_to_hybridization_rate
+      restrict_network_to_leaf_set
+      suppress_degree_two_nodes
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+

[…]
