### 2024-05-13 69c15bb
add docstrings for lgt and creating a network with cps
