### 2024-05-13 88637d6
fix and test newick input for hybridization problem class (#61)
* fix and test newick input for hybridization problem class

* changelog
