### 2024-05-13 4f6be0c
add docstrings for lgt and creating a network with cps (#60)
* add docstrings for lgt and creating a network with cps

* make docs and changelog
