+++ src/phylox/generators/lgt/base.py
@@ -15,28 +15,76 @@ from networkx.utils.decorators import np_random_state, py_random_state
 from phylox import DiNetwork
 
 
-def last_node(net):
+def _last_node(net):
+    """
+    The lgt generator names its nodes with integers, so the last node added
+    is the node represented with the maximum integer.
+
+    :param net: a phylox DiNetwork
+    :return: the node of the network with the highest integer associated to it.
+    """
     return max(net.nodes())
 
 
 def speciate(net, leaf):
-    l = last_node(net)
+    """
+    Add a new leaf to the network by a 'speciation event' at leaf `leaf` of the network.
+    This is done by adding two outgoing edges to the `leaf` node.
+
+    :param net: a phylox DiNetwork
+    :param leaf: leaf of the network `net`
+    :return: None
+    """
+    l = _last_node(net)
     net.add_edge(leaf, l + 1)
     net.add_edge(leaf, l + 2)
 
 
 def lgt(net, leaf1, leaf2):
-    net.add_edge(leaf1, leaf2, secondary=True)
-    l = last_node(net)
+    """
+    Adds a 'lateral gene transfer' to the network between leaves `leaf1` and 
+    `leaf2`. This is done by adding an outgoing edge to each of the leaf 
+    nodes and adding an edge between these nodes as well.
+
+    :param net: a phylox DiNetwork
+    :param leaf1: a leaf of the network `net`
+    :param leaf2: a leaf of the network `net`
+    :return: None
+    """    
+    net.add_edge(leaf1, leaf2)
+    l = _last_node(net)
     net.add_edge(leaf1, l + 1)
     net.add_edge(leaf2, l + 2)
 
 
 def leaves(net):
+    """
+    Returns a list of leaves of the network.
+
+    :param net: a phylox DiNetwork
+    :return: a list of leaf nodes of the network
+    """
     return [u for u in net.nodes() if net.out_degree(u) == 0]
+
+def reticulations(net):
+    """
+    Returns a list of reticulation nodes of the network.
+
+    :param net: a phylox DiNetwork
+    :return: a list of reticulation nodes of the network
+    """
+    return [v for v in net.nodes() if net.in_degree(v) == 2]
 
 
 def internal_blobs(net):
+    """
+    Returns a list of internal blobs of the network.
+    An internal blob is a biconnected component of 
+    the underlying undirected graph of the network, after removing the leaves.
+
+    :param net: a phylox DiNetwork
+    :return: a list of internal blobs of the network
+    """
     internal_nodes = set([u for u in net.nodes() if net.out_degree(u) > 0])
     blobs = list(nx.biconnected_components(nx.Graph(net)))
     blobs = [bl for bl in blobs if len(bl) > 2]
@@ -47,6 +95,13 @@ def internal_blobs(net):
 
 
 def compute_hash(net):
+    """
+    Returns a dictionary, mapping each internal node to its blob, 
+    and each leaf to the blob directly adjacent to it.
+
+    :param net: a phylox DiNetwork
+    :return: a dictionary mapping each leaf to an internal blob.
+    """
     mapping_blobs = {}
     blobs = internal_blobs(net)
     for blob in blobs:
@@ -61,6 +116,14 @@ def compute_hash(net):
 
 
 def internal_and_external_pairs(net):
+    """
+    Returns a list internal pairs, and a list of external pairs.
+    A pair of leaves is internal if they are adjacent to the same internal blob, 
+    the pair is external otherwise.
+    
+    :param net: a phylox DiNetwork
+    :return: a list internal pairs, and a list of external pairs
+    """
     lvs = leaves(net)
[…]
