+++ tests/generators/test_lgt.py
@@ -1,14 +1,14 @@
 import unittest
 
-from phylox.generators.lgt import generate_network_lgt
+from phylox.generators.lgt import generate_network_lgt, generate_network_lgt_conditional
 from phylox.isomorphism import is_isomorphic
 
 
 class TestLGTNetwork(unittest.TestCase):
     def test_one_leaf(self):
         network = generate_network_lgt(
-            n=1,
-            k=0,
+            leaves_goal=1,
+            retics_goal=0,
             wint=1,
             wext=1,
         )
@@ -17,8 +17,8 @@ class TestLGTNetwork(unittest.TestCase):
 
     def test_one_leaf_many_retics(self):
         network = generate_network_lgt(
-            n=1,
-            k=100,
+            leaves_goal=1,
+            retics_goal=100,
             wint=1,
             wext=1,
         )
@@ -27,8 +27,8 @@ class TestLGTNetwork(unittest.TestCase):
 
     def test_two_leaves(self):
         network = generate_network_lgt(
-            n=2,
-            k=0,
+            leaves_goal=2,
+            retics_goal=0,
             wint=1,
             wext=1,
         )
@@ -37,8 +37,8 @@ class TestLGTNetwork(unittest.TestCase):
 
     def test_one_reticulation(self):
         network = generate_network_lgt(
-            n=4,
-            k=1,
+            leaves_goal=4,
+            retics_goal=1,
             wint=1,
             wext=1,
         )
@@ -47,8 +47,8 @@ class TestLGTNetwork(unittest.TestCase):
 
     def test_larger_internal(self):
         network = generate_network_lgt(
-            n=100,
-            k=10,
+            leaves_goal=100,
+            retics_goal=10,
             wint=1,
             wext=0.001,
         )
@@ -57,8 +57,8 @@ class TestLGTNetwork(unittest.TestCase):
 
     def test_larger_external(self):
         network = generate_network_lgt(
-            n=100,
-            k=10,
+            leaves_goal=100,
+            retics_goal=10,
             wint=0.001,
             wext=1,
         )
@@ -67,22 +67,85 @@ class TestLGTNetwork(unittest.TestCase):
 
     def test_seed(self):
         network1 = generate_network_lgt(
-            n=100,
-            k=10,
+            leaves_goal=100,
+            retics_goal=10,
             wint=1,
             wext=1,
             seed=1,
         )
         network2 = generate_network_lgt(
-            n=100,
-            k=10,
+            leaves_goal=100,
+            retics_goal=10,
             wint=1,
             wext=1,
             seed=1,
         )
         network3 = generate_network_lgt(
-            n=100,
-            k=10,
+            leaves_goal=100,
+            retics_goal=10,
+            wint=1,
+            wext=1,
+            seed=2,
+        )
+        self.assertTrue(is_isomorphic(network1, network2))
+        self.assertFalse(is_isomorphic(network1, network3))
+
+
+class TestLGTNetworkConditional(unittest.TestCase):
+    def test_one_leaf(self):
+        network = generate_network_lgt_conditional(
+            n=1,
+            k=0,
+            prob_lgt=0.5,
+            wint=1,
+            wext=1,
+        )
+        self.assertEqual(len(network.leaves), 1)
+        self.assertEqual(len(network.edges), 1)
+
+    def test_two_leaves(self):
+        network = generate_network_lgt_conditional(
+            n=2,
+            k=0,
+            prob_lgt=0.5,
+            wint=1,
+            wext=1,
+        )
+        self.assertEqual(len(network.leaves), 2)
+        self.assertEqual(len(network.edges), 3)
+
+    def test_one_reticulation(self):
+        network = generate_network_lgt_conditional(
[…]
