+++ docs/source/_autosummary/phylox.generators.lgt.base.last_node.rst
@@ -1,6 +1,0 @@
-phylox.generators.lgt.base.last\_node
-=====================================
-
-.. currentmodule:: phylox.generators.lgt.base
-
-.. autofunction:: last_node
