### 2024-05-15 f79bb7a
add test, fix problem with cache and with large degree roots
