+++ src/phylox/cherrypicking/base.py
@@ -245,6 +245,7 @@ def reduce_pair(network, x, y, inplace=False, nodes_by_label=False):
         network.remove_edge(py, px)
         suppress_node(network, px)
         suppress_node(network, py)
+    network._clear_cached()
     return network, cherry_type
 
 
@@ -510,9 +511,6 @@ def cherry_height(network, x, y):
     float
         The height of the cherry (x,y) if it is a cherry, False otherwise
     """
-    print("cherry_height")
-    print(network.edges(data=True))
-
     if (not x in network.leaves) or (not y in network.leaves):
         return False
     px = network.parent(x)
@@ -533,16 +531,16 @@ class CherryPickingMixin:
     @classmethod
     def from_cherry_picking_sequence(cls, sequence, heights=None, label_leaves=True):
         """
-        Creates a PhyloX DiNetwork network from a cherry picking sequence, 
+        Creates a PhyloX DiNetwork network from a cherry picking sequence,
         and possibly a matching sequence of heights of the cherries.
 
         :param sequence: a cherry picking sequence (i.e., a list of 2-tuples)
-        :param heights: a list of positive floats with the same length as 
+        :param heights: a list of positive floats with the same length as
           `sequence`. If None, the heights will be set to consecutive integers
-        :param label_leaves: Bool, whether to label the leaves 
+        :param label_leaves: Bool, whether to label the leaves
           with the nodes/labels used in the sequence
         :return: a network.
-        """        
+        """
         network = cls()
         heights = heights or [[h, h] for h in range(1, len(sequence) + 1)]
         for pair, height in zip(reversed(sequence), reversed(heights)):
