+++ src/phylox/exceptions/__init__.py
@@ -1,3 +1,7 @@
+"""
+Custom exceptions for PhyloX
+"""
+
 from .invalidmovedefinitionexception import InvalidMoveDefinitionException
 from .invalidmoveexception import InvalidMoveException
 from .nosolutionexception import NoSolutionException
