+++ src/phylox/__init__.py
@@ -1,2 +1,1 @@
 from phylox.dinetwork import *
-from phylox.base import *
