+++ docs/source/_autosummary/phylox.constants.rst
@@ -1,4 +1,4 @@
-phylox.constants
+﻿phylox.constants
 ================
 
 .. automodule:: phylox.constants
