### 2024-05-20 90c982e
Improve navigability of docs (#64)
* better nav menu and some addtional links on docs homepage

* changelog
