+++ docs/source/_autosummary/phylox.newick_parser.extended_newick_to_dinetwork.rst
@@ -1,0 +1,6 @@
+phylox.newick\_parser.extended\_newick\_to\_dinetwork
+=====================================================
+
+.. currentmodule:: phylox.newick_parser
+
+.. autofunction:: extended_newick_to_dinetwork
