+++ docs/source/_autosummary/phylox.dinetwork.remove_unlabeled_leaves.rst
@@ -1,0 +1,6 @@
+phylox.dinetwork.remove\_unlabeled\_leaves
+==========================================
+
+.. currentmodule:: phylox.dinetwork
+
+.. autofunction:: remove_unlabeled_leaves
