+++ docs/source/_autosummary/phylox.parser.extended_newick_to_dinetwork.rst
@@ -1,6 +1,0 @@
-phylox.parser.extended\_newick\_to\_dinetwork
-=============================================
-
-.. currentmodule:: phylox.parser
-
-.. autofunction:: extended_newick_to_dinetwork
