+++ docs/source/_autosummary/phylox.classes.rst
@@ -1,4 +1,4 @@
-phylox.classes
+﻿phylox.classes
 ==============
 
 .. automodule:: phylox.classes
