+++ docs/source/_autosummary/phylox.rst
@@ -1,42 +1,0 @@
-﻿phylox
-======
-
-.. automodule:: phylox
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree:
-   :template: custom-module-template.rst
-   :recursive:
-
-   phylox.base
-   phylox.cherrypicking
-   phylox.classes
-   phylox.constants
-   phylox.dinetwork
-   phylox.exceptions
-   phylox.generators
-   phylox.isomorphism
-   phylox.networkproperties
-   phylox.parser
-   phylox.rearrangement
-
