### 2024-05-20 #6427 @RemieJanssen
@bgyori I have filled in the missing docstrings wherever I could find them, and I have improved (hopefully) the navigability of the docs: https://phylox.readthedocs.io/en/latest/

If you find anything else, please let me know :)
