### 2024-06-03 8422fc4
Add DOIs to joss paper bib (#66)
