+++ joss/paper.bib
@@ -6,6 +6,7 @@
   address = {Pasadena, CA USA},
   year =      {2008},
   editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
+  url = {https://www.osti.gov/biblio/960616}
 }
 
 @article{10.1093/bioinformatics/btp163,
@@ -31,6 +32,8 @@
   volume={13},
   pages={1--9},
   year={2012},
+  doi = {10.1186/1471-2105-13-209},
+  url = {https://doi.org/10.1186/1471-2105-13-209},
   publisher={Springer}
 }
 
@@ -42,14 +45,19 @@
   number={12},
   pages={1569--1571},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btq228},
+  url = {https://doi.org/10.1093/bioinformatics/btq228}
 }
 
 @phdthesis{janssen2021rearranging,
   title={Rearranging phylogenetic networks},
   author={Janssen, R},
   year={2021},
-  school={Ph. D. thesis, Delft University of Technology}
+  school={Ph. D. thesis, Delft University of Technology},
+  ISBN={978-94-6332-758-9},
+  url={https://doi.org/10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57},
+  doi={10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57}
 }
 
 @inproceedings{janssen2020linear,
@@ -58,18 +66,22 @@
   booktitle={International Conference on Algorithms for Computational Biology},
   pages={93--107},
   year={2020},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-42266-0_8},
+  url={https://doi.org/10.1007/978-3-030-42266-0_8}
 }
 
 @article{pons2019generation,
-  title={Generation of Level-$ k $ k LGT Networks},
+  title={Generation of Level-k LGT Networks},
   author={Pons, Joan Carles and Scornavacca, Celine and Cardona, Gabriel},
   journal={IEEE/ACM transactions on computational biology and bioinformatics},
   volume={17},
   number={1},
   pages={158--164},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TCBB.2019.2895344},
+  url={https://doi.org/10.1109/TCBB.2019.2895344}
 }
 
 @article{zhang2018bayesian,
@@ -80,7 +92,9 @@
   number={2},
   pages={504--517},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/molbev/msx307},
+  url={https://doi.org/10.1093/molbev/msx307}
 }
 
 @inproceedings{janssen2020combining,
@@ -89,7 +103,9 @@
   booktitle={International Conference on Algorithms for Computational Biology},
   pages={77--92},
   year={2020},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-42266-0_7},
+  url={https://doi.org/10.1007/978-3-030-42266-0_7}
 }
 
 @article{kong2022classes,
@@ -100,7 +116,9 @@
   number={6},
   pages={47},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00285-022-01746-y},
+  url={https://doi.org/10.1007/s00285-022-01746-y}
 }
 
 @ARTICLE{4407681,
@@ -112,14 +130,19 @@
   number={4},
   pages={552-569},
[…]
