+++ joss/paper.md
@@ -135,7 +135,7 @@ with open("sampled_networks.nwk", "w") as f:
 
 For this sampler to work correctly, the space of networks that is sampled from needs to be connected. That is, it has to be possible to transform each network into each other network in the space using the selected rearrangement moves. In the example above, this means that the space of orchard network with 10 leaves and 5 reticulations needs to be connected under tail moves and head moves (i.e. rSPR moves).
 
-This is something the user needs to check or prove themselves, as it is not viable to check this computationally. Fortunately, such connectivity results have been studied in detail [@klawitter2020spaces; @thesis_janssen; @van2022orchard; @ERDOS2021205]. For example, the result needed to prove that this example is correct can be found in [@van2022orchard].
+This is something the user needs to check or prove themselves, as it is not viable to check this computationally. Fortunately, such connectivity results have been studied in detail [@klawitter2020spaces; @janssen2021rearranging; @van2022orchard; @ERDOS2021205]. For example, the result needed to prove that this example is correct can be found in [@van2022orchard].
 
 ## Comparing networks
 Based on all the properties above, PhyloX provides a toolkit to compare networks. For example, it can be used to determine whether two networks are [isomorphic](https://phylox.readthedocs.io/en/v1.0.3/_autosummary/phylox.isomorphism.base.is_isomorphic.html#phylox.isomorphism.base.is_isomorphic) (i.e., the same); whether they have the same [properties](https://phylox.readthedocs.io/en/v1.0.3/_autosummary/phylox.networkproperties.html): level, number of blobs, reticulation number, and number of (reticulated) cherries; whether one is [contained](https://phylox.readthedocs.io/en/v1.0.3/_autosummary/phylox.cherrypicking.tree_child_sequences.html#phylox.cherrypicking.tree_child_sequences.tree_child_network_contains) in the other, if both are tree-child; and whether they are similar with respect to a [rearangement distance](https://phylox.readthedocs.io/en/v1.0.3/_autosummary/phylox.rearrangement.exact_distance.html#module-phylox.rearrangement.exact_distance).
