### 2024-06-03 297a7cb
small fixes in joss text (#67)
