### 2024-06-03 8988038
add joss and docs badges
