### 2024-06-03 #6427 @RemieJanssen
Thanks @bgyori and @abhishektiwari for the helpful reviews, and @fboehm for the smooth process!
I have addressed the last PR from bgyori and fixed a few typos myself as well.
This must mean that we can finish up this phase, is that correct?

The last thing I'd like to fix is some metadata.
I see the version is still 1.0.3 at the top of this thread.
The version the paper corresponds to should be updated I think, is that still possible, and how do I do that?
