### 2024-06-16 #6427 @bgyori
Yes, meets all review criteria
