### 2024-07-15 5edc7f2
prepare for JOSS release (#69)
Add .zenodo.json file and fix typos
