+++ joss/paper.md
@@ -45,14 +45,14 @@ Kong et al. [@kong2022classes] gives a good overview of most well-known classes 
 ## Cherry-picking
 A basic structure in any network or tree is the *cherry*, a pair of leaves with a common parent. A modified version often found in phylogenetic networks is the *reticulated cherry*, an ordered pair of leaves $(x, y)$ which are related through the three edges $(p_x, x)$, $(p_y, y)$, and $(p_y, p_x)$.
 
-A common modification to a phylogenetic network is to *pick* or ( or *reduce*) a cherry or reticulated cherry. To pick a cherry $(x,y)$, one removes the leaf $x$ from the network together with its incoming edge, and then suppresses the resulting degree 2 node if the shared parent of $x$ and $y$ had out-degree 2. *Suppressing* a degree 2 node consists in removing the node and its two incident edges, and replacing them all with one new edge. To pick a reticulated cherry $(x,y)$, one removes the edge $(p_y, p_x)$, and supresses all resulting degree 2 nodes. The reverse action of picking a cherry, is called *adding a cherry*.
+A common modification to a phylogenetic network is to *pick* or ( or *reduce*) a cherry or reticulated cherry. To pick a cherry $(x,y)$, one removes the leaf $x$ from the network together with its incoming edge, and then suppresses the resulting degree 2 node if the shared parent of $x$ and $y$ had out-degree 2. *Suppressing* a degree 2 node consists in removing the node and its two incident edges, and replacing them all with one new edge. To pick a reticulated cherry $(x,y)$, one removes the edge $(p_y, p_x)$, and suppresses all resulting degree 2 nodes. The reverse action of picking a cherry, is called *adding a cherry*.
 
 These modifications are used in computational tools, for example to reconstruct networks from ancestral profiles [@ERDOS201933; @bai2021defining; @cardona2024comparison], to check whether one tree-child network is contained in another [@janssen2021cherry], or to combine multiple trees into one network [@linz2019attaching; @van2022practical]. Their versatile use has also lead to the introduction of the class of *orchard networks* [@janssen2021cherry; @ERDOS201933]. This class contains all networks that can be reduced to a single leaf using cherry-picking operations. Networks from this class can be interpreted as trees with horizontal gene transfer arcs [@van2022orchard].
 
 ## Rearranging networks
 For phylogenetic inference problems, it is often necessary to use heuristics that search through a space of networks. Such a space of networks takes the shape of a graph, whose objects are all networks with a common set of leaf labels (the sampled taxa), and sometimes also a set number of reticulations. The edges of the graph correspond to small changes made to a network: there is an edge between two networks if one can make a modification to one of the networks to arrive at the second network.
 
-The modifications that are allowed are well-defined as types of *rearrangement* opterations. These operations can be *horizontal*, keeping the reticulation number the same, or *vertical*, changing the reticulation number. Most horizontal operations are a variation on or restriction of the *rooted subtree prune and regraft* (*rSPR*) operation [@bordewich2017lost; @gambette2017rearrangement].
+The modifications that are allowed are well-defined as types of *rearrangement* operations. These operations can be *horizontal*, keeping the reticulation number the same, or *vertical*, changing the reticulation number. Most horizontal operations are a variation on or restriction of the *rooted subtree prune and regraft* (*rSPR*) operation [@bordewich2017lost; @gambette2017rearrangement].
 
 The names for vertical moves have not been standardized, but they generally do the same. A vertical move that removes a reticulation removes an incoming edge of a reticulation node, and then suppresses the resulting degree 2 nodes. A vertical move that adds a reticulation does the reverse: it *subdivides* two edges of the network, and adds a new edge between the two new degree 2 nodes (e.g. [@bordewich2017lost]).
 
[…]
