+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Types of changes:
 - Fixed for any bug fixes.
 - Security in case of vulnerabilities.
 
-## [1.0.6] - (Unreleased)
+## [1.1.0] - (Unreleased)
 
  - Improve navigability of docs. (#64)
 
