### 2024-07-15 #6427 @RemieJanssen
No real questions, I have checked everything, and it all looks correct to me now.
I am intersted though: I have set the zenodo metadata (author and license) using the .zenodo.json file, is that the recommended way? I also saw there is the option of adding a CITATION.cff file, but I read that overwrites all the harvested GitHub metadata in Zenodo.

Version number: v1.1.0
Zenodo DOI: https://doi.org/10.5281/zenodo.12742473

Thank you @bgyori and @abhishektiwari for all your helpful comments, I really do think the paper as well as the package have improved a lot due to your insights! And a lot of thanks to @fboehm for leading me through this process. This was my first time working with JOSS and a very open review process. You made it into a quite pleasant ride :)
