### 2024-07-19 #6427 @fboehm
@RemieJanssen - thank you for the updates!

Let me check on your question about setting zenodo metadata. If you have a moment, can you point me to where you read concerns about using a CITATION.cff file? I've never encountered your questions before, so I want to give a complete answer.

Thanks again!
