### 2024-07-21 #6427 @RemieJanssen
I'm not sure my concerns are valid, and I think my solution using the `.zenodo.json` file is fine. So there's no real need to dive into it, unless you want to of course!

I think I decided to not use the CITATION.cff file after reading these comments in the same thread:
https://github.com/citation-file-format/citation-file-format/issues/374#issuecomment-1689701707
https://github.com/citation-file-format/citation-file-format/issues/374#issuecomment-2150918724
