### 2024-08-01 #6427 @fboehm
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1093/bioinformatics/btp163 is OK
> - 10.1186/1471-2105-13-209 is OK
> - 10.1093/bioinformatics/btq228 is OK
> - 10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57 is OK
> - 10.1007/978-3-030-42266-0_8 is OK
> - 10.1109/TCBB.2019.2895344 is OK
> - 10.1093/molbev/msx307 is OK
> - 10.1007/978-3-030-42266-0_7 is OK
> - 10.1007/s00285-022-01746-y is OK
> - 10.1109/TCBB.2007.70270 is OK
> - 10.1017/CBO9780511974076 is OK
> - 10.1016/j.mbs.2019.04.009 is OK
> - 10.1016/j.mbs.2021.108537 is OK
> - 10.1109/TCBB.2024.3361390 is OK
> - 10.1016/j.tcs.2020.12.031 is OK
> - 10.1007/s00453-021-00914-8 is OK
> - 10.1016/j.aam.2019.01.004 is OK
> - 10.1007/s11538-022-01037-z is OK
> - 10.1007/s00285-021-01662-7 is OK
> - 10.1142/S0219720021400126 is OK
> - 10.1186/1471-2105-9-532 is OK
> - 10.1093/sysbio/syy015 is OK
> - 10.1111/mec.13544 is OK
> - 10.1073/pnas.1407950111 is OK
> - 10.1016/j.jtbi.2017.03.032 is OK
> - 10.1371/journal.pcbi.1005611 is OK
> - 10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1 is OK
> - 10.1186/s13015-023-00233-3 is OK
> - 10.1007/s11538-022-01037-z is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Exploring Network Structure, Dynamics, and Functio...
> - No DOI given, and none found for title: Rich newick format
> - No DOI given, and none found for title: Maximum likelihood inference of reticulate evoluti...
> - No DOI given, and none found for title: Spaces of phylogenetic networks
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.dam.2021.02.016 is INVALID because of 'https://doi.org/' prefix
> ```

@RemieJanssen - could you fix the bib file so that these issues resolve? I think one problem is with the prefix for one of dois. It also looks like several entries might be missing dois - can you add them?
