### 2024-08-02 #6427 @RemieJanssen
@fboehm Oh, those must have slipped through!

I was able to fix:
 - No DOI given, and none found for title: Maximum likelihood inference of reticulate evoluti...
- https://doi.org/10.1016/j.dam.2021.02.016 is INVALID because of 'https://doi.org/' prefix

The other ones do not have a DOI:
- No DOI given, and none found for title: Exploring Network Structure, Dynamics, and Functio...

This one is most surprising, as it seems to be a conference article https://www.osti.gov/biblio/960616. It is the article for the networkx python package and I used their own recommended bibtex to cite them though... https://networkx.org/documentation/stable/index.html

- No DOI given, and none found for title: Rich newick format

This is just a website. But it is the best source I could find for this particular file format.

- No DOI given, and none found for title: Spaces of phylogenetic networks

This is a PhD thesis. I could not find a DOI, but the handle/url is given in the reference.
