### 2024-08-06 #6427 @fboehm
@RemieJanssen - did you push the bib file fixes to the main branch of the repository? It seems that I'm still seeing the error about the https prefix.
