### 2024-08-06 #6427 @fboehm
I think that this line needs to be fixed: https://github.com/RemieJanssen/PhyloX/blob/5edc7f25d3eb5f071bcdd3c620048153af785892/joss/paper.bib#L398
