### 2024-08-06 4e42cd1
add and fix dois (#70)
