+++ joss/paper.bib
@@ -313,8 +313,8 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
   pages={16448--16453},
   year={2014},
   publisher={National Acad Sciences},
-  doi={},
-  url={}
+  doi={10.1073/pnas.1407950111},
+  url={https://doi.org/10.1073/pnas.1407950111}
 }
 
 @article{wen2016bayesian,
@@ -395,7 +395,7 @@ volume = {294},
 pages = {205-213},
 year = {2021},
 issn = {0166-218X},
-doi = {https://doi.org/10.1016/j.dam.2021.02.016},
+doi = {10.1016/j.dam.2021.02.016},
 url = {https://www.sciencedirect.com/science/article/pii/S0166218X21000597},
 author = {Péter L. Erdős and Andrew Francis and Tamás Róbert Mezei},
 keywords = {Nearest-neighbour interchange, Rooted NNI moves, Distance-1 tail moves, Rooted phylogenetic network, Directed acyclic graph, Tree-based network},
