### 2024-08-31 #6427 @RemieJanssen
@fboehm Could you give me an update? If you are waiting for me to fix the last three DOIs I must disappoint you, as they are simply not available. I could fix that by simply removing the references, or replacing them with a footnote. That feels like a decent solution for the website 'Rich newick format'. For the other two, that does not sit entirely right with me. As they are published scientific works (a PhD thesis and a conference article).
