### 2024-08-31 #6427 @fboehm
both archive & software repository use the same license - bsd3
