### 2024-09-01 #6427 @fboehm
@RemieJanssen - In reading the generated pdf of the manuscript, I have a few small requests:

1. line 41 of the pdf says: "Kong et al. (Kong et al., 2022)" - is it possible to change this so that the text "Kong et al." appears only once here? perhaps the entire phrase "Kong et al., 2022" can still be the hyperlink.

2. In at least two places, a 2021 work is referenced as "Remie Janssen & Murakami, 2021" (lines 60 & 63 of pdf). Can you fix this so that only last (family) names are used in the text? Same with tehe 2021 article referenced on line 94 of pdf: "Remie Janssen & Liu"; line 157 of pdf (two refs)

Otherwise, I think that the pdf text looks good! Thanks again!
