+++ joss/paper.bib
@@ -62,7 +62,7 @@
 
 @inproceedings{janssen2020linear,
   title={Linear time algorithm for tree-child network containment},
-  author={Janssen, R and Murakami, Yukihiro},
+  author={Janssen, Remie and Murakami, Yukihiro},
   booktitle={International Conference on Algorithms for Computational Biology},
   pages={93--107},
   year={2020},
