### 2024-09-11 #6427 @RemieJanssen
@fboehm I'm sorry, I must have missed the notifications for your previous message... I do understand your comments and will try to fix them as best I can.

The one for Kong et al. I can definitely fix :)

The one about the citation style may lead me into a rabbit hole:
 - https://github.com/openjournals/joss/issues/400
 - https://github.com/openjournals/joss/issues/230
 - https://github.com/jgm/pandoc-citeproc/issues/184

I haven't got it to look nice yet, but I will continue trying for a bit.
