+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 name = "phylox"
 version = "1.1.0"
 authors = [
-  { name="Remie Janssen", email="remiejanssen92@gmail.com" },
+  { name="Henk Harmsen", email="remiejanssen92@gmail.com" },
 ]
 description = "A package with tools for constructing, manipulating, and analyzing phylogenetic networks."
 readme = "README.md"
