+++ joss/paper.bib
@@ -62,7 +62,7 @@
 
 @inproceedings{janssen2020linear,
   title={Linear time algorithm for tree-child network containment},
-  author={Janssen, Remie and Murakami, Yukihiro},
+  author={Murakami, Yukihiro and Janssen, Remie},
   booktitle={International Conference on Algorithms for Computational Biology},
   pages={93--107},
   year={2020},
@@ -184,7 +184,7 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
 
 @article{janssen2021cherry,
   title={On cherry-picking and network containment},
-  author={Janssen, Remie and Murakami, Yukihiro},
+  author={Murakami, Yukihiro and Janssen, Remie},
   journal={Theoretical Computer Science},
   volume={856},
   pages={121--150},
