+++ joss/paper.bib
@@ -246,7 +246,7 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
 
 @article{janssen2021comparing,
   title={Comparing the topology of phylogenetic network generators},
-  author={Janssen, Remie and Liu, Pengyu},
+  author={Liu, Pengyu and Janssen, Remie},
   journal={Journal of bioinformatics and computational biology},
   volume={19},
   number={06},
