### 2024-09-11 fe724e6
suggest initial to latex
