+++ joss/paper.md
@@ -7,7 +7,7 @@ tags:
   - Graph theory
 authors:
   - name: Remie Janssen
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0002-5192-1470
     affiliation: 1
 affiliations:
  - name: National Institute for Public Health and the Environment, Bioinformatics and Computing group, Bilthoven, The Netherlands
