### 2024-09-11 ad29fc0
fix strange references in text with first name Remie (#71)
* try to fix by changing my name in md

* use newer version of upload artifact

* initial instead of name

* initial instead of name

* change csl to apa

* change csl to chicago

* fix kong

* fix kong

* change csl

* see if changing the author of the paper fixes anything

* It does not... try changing owner of the package

* swap order of authors in papers

* swap back

* swap authors of liu paper

* try changing author again

* suggest initial to latex

* change back initial thing

* be tricksy

* put back orcid id

---------

Co-a
[…]
