### 2024-09-11 #6427 @RemieJanssen
I managed to fix it, though it's kind of hacky in the bib file...
https://github.com/RemieJanssen/PhyloX/pull/71

For anyone who runs into the same problem, this is how I did it:
```
author={ and Janssen, Remie and Murakami, Yukihiro},
```
