### 2024-09-11 #6427 @RemieJanssen
@fboehm The strange references should be fixed now. They look good in my drafts, but it merits checking if the paper also comes out looking good when it's compiled here by the JOSS scripts...
