### 2024-09-26 #6427 @fboehm
@RemieJanssen - I can't recall - did we resolve the issue with the reference for your phd thesis & the way that the author(s) get listed? I'm still seeing it, in the pdf linked above (about line 255 in the pdf), as:

Janssen, R and .  (2021)

I think we want:

Janssen, R.  (2021)

for your phd thesis.

Can you try editing line 55 of the bib file? https://github.com/RemieJanssen/PhyloX/blob/ad29fc08b9c4dbd8c3f7efd5a057193454841aed/joss/paper.bib#L55

maybe try replacing your current line 55 with:

author={Janssen, Remie},

thanks again!
