### 2024-09-27 f431e5a
fix reference phd thesis
