+++ joss/paper.bib
@@ -52,7 +52,7 @@
 
 @phdthesis{janssen2021rearranging,
   title={Rearranging phylogenetic networks},
-  author={ and Janssen, Remie},
+  author={Janssen, Remie},
   year={2021},
   school={Ph. D. thesis, Delft University of Technology},
   ISBN={978-94-6332-758-9},
