### 2024-09-27 8dff907
Fix references Remie Janssen (#72)
* fix reference phd thesis

* remove all and before my name
