+++ joss/paper.bib
@@ -52,7 +52,7 @@
 
 @phdthesis{janssen2021rearranging,
   title={Rearranging phylogenetic networks},
-  author={ and Janssen, Remie},
+  author={Janssen, Remie},
   year={2021},
   school={Ph. D. thesis, Delft University of Technology},
   ISBN={978-94-6332-758-9},
@@ -62,7 +62,7 @@
 
 @inproceedings{janssen2020linear,
   title={Linear time algorithm for tree-child network containment},
-  author={ and Janssen, Remie and Murakami, Yukihiro},
+  author={Janssen, Remie and Murakami, Yukihiro},
   booktitle={International Conference on Algorithms for Computational Biology},
   pages={93--107},
   year={2020},
@@ -99,7 +99,7 @@
 
 @inproceedings{janssen2020combining,
   title={Combining networks using cherry picking sequences},
-  author={ and Janssen, Remie and Jones, Mark and Murakami, Yukihiro},
+  author={Janssen, Remie and Jones, Mark and Murakami, Yukihiro},
   booktitle={International Conference on Algorithms for Computational Biology},
   pages={77--92},
   year={2020},
@@ -184,7 +184,7 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
 
 @article{janssen2021cherry,
   title={On cherry-picking and network containment},
-  author={ and Janssen, Remie and Murakami, Yukihiro},
+  author={Janssen, Remie and Murakami, Yukihiro},
   journal={Theoretical Computer Science},
   volume={856},
   pages={121--150},
@@ -246,7 +246,7 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
 
 @article{janssen2021comparing,
   title={Comparing the topology of phylogenetic network generators},
-  author={ and Janssen, Remie and Liu, Pengyu},
+  author={Janssen, Remie and Liu, Pengyu},
   journal={Journal of bioinformatics and computational biology},
   volume={19},
   number={06},
