### 2024-09-27 #6427 @RemieJanssen
Looks like it is finally fixed, and without a dumb trick! :) Looks like the original issue was that I had both "Janssen, R" and "Janssen, Remie" in my bib file, so it wanted to resolve it by spelling out the first name. My bad...
