### 2024-09-27 #6427 @fboehm
Can you add the title to the bibliography entry for Julien, et al.
I think that this can be accomplished by changing this line of the bib file:

https://github.com/RemieJanssen/PhyloX/blob/8dff90726ea1a12463052a97925142d2e4543b50/joss/paper.bib#L366

from

@misc{https://doi.org/10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1,

to

@article{https://doi.org/10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1,

Can you check if this works? And push an update to github if it does?

Thanks again, @RemieJanssen
