### 2024-09-27 cebae0f
fix reference phdthesis remie janssen
