+++ joss/paper.bib
@@ -54,7 +54,7 @@
   title={Rearranging phylogenetic networks},
   author={Janssen, Remie},
   year={2021},
-  school={Ph. D. thesis, Delft University of Technology},
+  school={Delft University of Technology},
   ISBN={978-94-6332-758-9},
   url={https://doi.org/10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57},
   doi={10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57}
