### 2024-09-27 d45b0e3
Fix references: Julien et al and PhD thesis Janssen (#73)
* fix title julien et al

* fix reference phdthesis remie janssen
