+++ joss/paper.bib
@@ -54,7 +54,7 @@
   title={Rearranging phylogenetic networks},
   author={Janssen, Remie},
   year={2021},
-  school={Ph. D. thesis, Delft University of Technology},
+  school={Delft University of Technology},
   ISBN={978-94-6332-758-9},
   url={https://doi.org/10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57},
   doi={10.4233/uuid:1b713961-4e6d-4bb5-a7d0-37279084ee57}
@@ -364,11 +364,11 @@ abstract = {Rooted phylogenetic networks provide an explicit representation of t
 }
 
 @misc{https://doi.org/10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1,
+  title = {Source code for the paper "Constructing Phylogenetic Networks via Cherry Picking and Machine Learning"},
   doi = {10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1},
   url = {https://doi.org/10.4121/c679cd3c-0815-4021-a727-bcb8b9174b27.v1},
   author = {Julien, Esther and Bernardini, Giulia and Stougie, Leen and van Iersel, Leo},
   keywords = {Phylogenetics, Hybridization, Cherry Picking, Machine Learning, Heuristic},
-  title = {Source code for the paper &#34;Constructing Phylogenetic Networks via Cherry Picking and Machine Learning&#34;},
   publisher = {4TU.ResearchData},
   year = {2023},
   copyright = {MIT}
