### 2024-09-27 #6427 @RemieJanssen
I've fixed it, there was an issue with the quotes in the title it seems.
I went through the bibliography again to see if anything else stood out, and caught one last (I hope) little thing.

Thanks for being so thorough and patient!
