### 2024-11-01 #6427 @arfon
@RemieJanssen – it looks like there is a duplicate BibTeX entry here. Could you check your references and see if you can spot the error?
