+++ joss/paper.bib
@@ -401,16 +401,3 @@ author = {Péter L. Erdős and Andrew Francis and Tamás Róbert Mezei},
 keywords = {Nearest-neighbour interchange, Rooted NNI moves, Distance-1 tail moves, Rooted phylogenetic network, Directed acyclic graph, Tree-based network},
 abstract = {We show that any two tree-based rooted binary phylogenetic networks (with the same number of reticulation nodes and identical degrees at the root) are connected by at most (4k+8)⋅|X|+34k+|X|⋅⌈log2|X|⌉ rooted nearest-neighbour interchange (rNNI) moves. As a corollary of the proof we obtain that the diameter is at most a 2k additive term larger for distance-1 tail moves.}
 }
-
-@article{van2022orchard,
-  title={Orchard networks are trees with additional horizontal arcs},
-  author={van Iersel, Leo and Janssen, Remie and Jones, Mark and Murakami, Yukihiro},
-  journal={Bulletin of Mathematical Biology},
-  volume={84},
-  number={8},
-  pages={76},
-  year={2022},
-  publisher={Springer},
-  doi={10.1007/s11538-022-01037-z},
-  url={https://doi.org/10.1007/s11538-022-01037-z}
-}
