### 2024-11-02 c094d57
add doi for networkx reference (#75)
