+++ joss/paper.bib
@@ -6,7 +6,8 @@
   address = {Pasadena, CA USA},
   year =      {2008},
   editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
-  url = {https://www.osti.gov/biblio/960616}
+  url = {https://www.osti.gov/biblio/960616},
+  doi = {10.25080/TCWV9851}
 }
 
 @article{10.1093/bioinformatics/btp163,
