### 2024-11-02 #6427 @RemieJanssen
@arfon I removed the duplicate.

I noticed it now also gave a suggestion for the NetworkX paper DOI (which I could not find myself). So I checked it and added it, too. I wonder why it did not suggest this before...
