### 2024-11-15 #6427 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
