### 2024-11-15 #6427 @Kevin-Mattheus-Moerman
@RemieJanssen as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Here are some final checks:

- [x] Software license tag listed here matches a tagged release

- [ ] Please ensure the archive listed title matches the paper title. There is no need to create a new archive, you can simple edit the existing one.
- [ ] Minor note on language/text. The following are examples of British English: `analysing, labelling, focussing`, whereas the following are examples of American English occurring in the text `characterized, standardized, initialized, utilized, standardizes, hybridization, initialization`. I would recommend that you chose one and update the text to be consistent.
