+++ joss/paper.md
@@ -2,7 +2,7 @@
 title: 'PhyloX: A Python package for complete phylogenetic network workflows'
 tags:
   - Python
-  - Bio-informatics
+  - Bioinformatics
   - Phylogenetics
   - Graph theory
 authors:
@@ -18,13 +18,13 @@ bibliography: paper.bib
 
 # Summary
 
-PhyloX is a Python package with tools for generating, manipulating, and analyzing phylogenetic networks. It uses the NetworkX package [@SciPyProceedings_11] for basic graph operations. This has the added benefit that the powerful graph tools from NetworkX can be used directly on the phylogenetic networks as well. The aim of the package is to be of general use to phylogenetic network researchers, with a current focus on I/O, random generation of networks, cherry-picking methods, rearrangement operations, and the identification of classes and properties of networks.
+PhyloX is a Python package with tools for generating, manipulating, and analysing phylogenetic networks. It uses the NetworkX package [@SciPyProceedings_11] for basic graph operations. This has the added benefit that the powerful graph tools from NetworkX can be used directly on the phylogenetic networks as well. The aim of the package is to be of general use to phylogenetic network researchers, with a current focus on I/O, random generation of networks, cherry-picking methods, rearrangement operations, and the identification of classes and properties of networks.
 
 
 
 # Phylogenetic networks
 
-In the study of the evolutionary history of biological species and languages, it is common to represent putative histories using graphs. Traditionally, at least in biology, these graphs are most often trees, such as the well known tree drawn by Charles Darwin in one of his notebooks. A tree like this is called a phylogenetic tree. In some cases, the evolutionary history includes complex processes like horizontal gene transfer and hybridization. These processes cause a *reticulate* (i.e. network-like) structure in the evolutionary history, which requires phylogenetic networks to be used for representing the evolutionary histories.
+In the study of the evolutionary history of biological species and languages, it is common to represent putative histories using graphs. Traditionally, at least in biology, these graphs are most often trees, such as the well-known tree drawn by Charles Darwin in one of his notebooks. A tree like this is called a phylogenetic tree. In some cases, the evolutionary history includes complex processes like horizontal gene transfer and hybridisation. These processes cause a *reticulate* (i.e., network-like) structure in the evolutionary history, which requires phylogenetic networks to be used for representing the evolutionary histories.
 
 A *directed phylogenetic network* (e.g., [@huson2010phylogenetic]) is a directed acyclic graph with four types of nodes:
  - a *root*: an in-degree 0, out-degree 1 node,
@@ -35,45 +35,45 @@ A *directed phylogenetic network* (e.g., [@huson2010phylogenetic]) is a directed
 A network is *binary* if each reticulation node has in-degree 2, and each tree node has in-degree 2. An *undirected phylogenetic network* is the underlying undirected graph of a directed phylogenetic network, retaining the labelling of the leaf nodes.
 
 ## Network properties
-When analysing or comparing phylogenetic networks or phylogenetic network methods, it can be helpful to extract some (numerical) parameters from the networks. Some of the most used properties are the *reticulation number* (the number of reticulation nodes in a binary network), the number of *blobs* (biconnected components of the network), and the *level* (the maximum reticulation number among all blobs of the network). Of course, the list of studied properties and parameters is much longer, including for example the recently introduced $B_2$-balance index of the network [@franccois2021revisiting].
[…]
