### 2024-11-24 #6427 @RemieJanssen
@Kevin-Mattheus-Moerman
I fixed the spelling as best I could https://github.com/RemieJanssen/PhyloX/pull/76
Can you explain to me the other point? Must the title of the zenodo record match exactly to the paper's title, or is there some other field where it should match (e.g., a 'related works' field with relation 'is published in')?
