### 2024-11-24 #6427 @Kevin-Mattheus-Moerman
@RemieJanssen yes please edit the archive title field so the title will match the paper title.
