### 2024-11-28 #6427 @RemieJanssen
> 🚨🚨🚨 **THIS IS NOT A DRILL, YOU HAVE JUST ACCEPTED A PAPER INTO JOSS!** 🚨🚨🚨
>
> Here's what you must now do:
>
> 0. Check final PDF and Crossref metadata that was deposited 👉 [Creating pull request for 10.21105.joss.06427 joss-papers#6180](https://github.com/openjournals/joss-papers/pull/6180)
> 1. Wait five minutes, then verify that the paper DOI resolves https://doi.org/10.21105/joss.06427
> 2. If everything looks good, then close this review issue.
> 3. Party like you just published a paper! 🎉🌈🦄💃👻🤘
>
> Any issues? Notify your editorial technical team...

@Kevin-Mattheus-Moerman Was this directed at me? Either way, I have checked the paper behind the DOI link, and it looks good to me.

Thanks everyone!
